+++ packages/patrol/example/integration_test/swipe_test.dart
@@ -15,15 +15,11 @@ void main() {
   patrol('scrolls the Settings app', ($) async {
     await createApp($);
 
-    final window = $.tester.binding.window;
-    final width = window.physicalSize.width;
-    final height = window.physicalSize.height;
-
     await $.native.openApp(appId: appId);
-
-    final from = Offset((width / 2) / width, (height / 2) / height);
-    final to = Offset((width / 2) / width, 100 / height);
-
-    await $.native.swipe(from: from, to: to);
+    await $.native.swipe(
+      from: Offset(0.5, 0.8),
+      to: Offset(0.5, 0.2),
+      appId: appId,
+    );
   });
 }