+++ packages/patrol/ios/Classes/AutomatorServer/Automator.swift
@@ -198,6 +198,17 @@
       sleepTask(timeInSeconds: 1)
     }
 
+    func swipe(from start: CGVector, to end: CGVector, inApp bundleId: String) throws {
+      try runAction("swiping from \(start) to \(end) in app \(bundleId)") {
+        let app = try self.getApp(withBundleId: bundleId)
+
+        let startCoordinate = app.coordinate(
+          withNormalizedOffset: CGVector(dx: start.dx, dy: start.dy))
+        let endCoordinate = app.coordinate(withNormalizedOffset: CGVector(dx: end.dx, dy: end.dy))
+        startCoordinate.press(forDuration: 0.1, thenDragTo: endCoordinate)
+      }
+    }
+
     func waitUntilVisible(onText text: String, inApp bundleId: String) throws {
       try runAction(
         "waiting until view with text \(format: text) in app \(bundleId) becomes visible"