+++ packages/patrol/ios/Classes/PatrolIntegrationTestRunner.h
@@ -1,9 +1,9 @@
 // This file is a one giant macro to make the setup as easy as possible for the developer.
 // To edit it:
-//  1. Remove the trailing backslashes: $ sed 's/\\$//' ios/Classes/PatrolIntegrationTestRunner.h
+//  1. Remove the trailing backslashes: $ sed 's/\\$//' PatrolIntegrationTestRunner.h
 //  2. Paste its contents into the RunnerUITests.m in the RunnerUITests target
 //  3. Make the changes, make sure it works
-//  4. Re-add trailing backslashes: $ sed 's/$/\\/' ios/Classes/PatrolIntegrationTestRunner.h
+//  4. Re-add trailing backslashes: $ sed 's/$/\\/' RunnerUITests.m
 //  5. Copy the contents from RunnerUITests.m back here
 //  6. Go back to using a macro in RunnerTests.m
 
@@ -16,8 +16,20 @@
   @implementation __test_class                                                                                      \
                                                                                                                     \
   +(NSArray<NSInvocation *> *)testInvocations {                                                                     \
+    __block NSMutableDictionary<NSString *, NSNumber *> *callbacksState = NULL;                                     \
+    __block NSArray<NSString *> *dartTests = NULL;                                                                  \
+    __block BOOL appReady = NO;                                                                                     \
+                                                                                                                    \
     /* Start native automation server */                                                                            \
-    PatrolServer *server = [[PatrolServer alloc] init];                                                             \
+    PatrolServer *server = [[PatrolServer alloc]                                                                    \
+        initWithOnAppReadyCallback:^{                                                                               \
+          appReady = YES;                                                                                           \
+        }                                                                                                           \
+        onDartLifecycleCallbackExecuted:^(NSString *_Nonnull callbackName) {                                        \
+          /* callbacksState dictionary will have been already initialized when this callback is executed */         \
+          NSLog(@"onLifecycleCallbackExecuted for %@", callbackName);                                               \
+          [callbacksState setObject:@YES forKey:callbackName];                                                      \
+        }];                                                                                                         \
                                                                                                                     \
     NSError *_Nullable __autoreleasing *_Nullable err = NULL;                                                       \
     [server startAndReturnError:err];                                                                               \
@@ -35,15 +47,45 @@
       [systemAlerts.buttons[@"Allow"] tap];                                                                         \
     }                                                                                                               \
                                                                                                                     \
-    /* Run the app for the first time to gather Dart tests */                                                       \
-    [[[XCUIApplication alloc] init] launch];                                                                        \
+    NSLog(@"Performing initial run of the app...");                                                                 \
                                                                                                                     \
-    /* Spin the runloop waiting until the app reports that it is ready to report Dart tests */                      \
-    while (!server.appReady) {                                                                                      \
+    /* MARK: Start initial run */                                                                                   \
+    /* Run the app for the first time to gather Dart tests */                                                       \
+    XCUIApplication *app = [[XCUIApplication alloc] init];                                                          \
+    NSDictionary *args = @{@"PATROL_INITIAL_RUN" : @"true"};                                                        \
+    app.launchEnvironment = args;                                                                                   \
+    [app launch];                                                                                                   \
+                                                                                                                    \
+    NSLog(@"Waiting for PatrolAppService (running in the app under test) to report its readiness...");              \
+                                                                                                                    \
+    /* Spin the runloop waiting until the app reports that PatrolAppService is up */                                \
+    while (!appReady) {                                                                                             \
       [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                            \
     }                                                                                                               \
                                                                                                                     \
-    __block NSArray<NSString *> *dartTests = NULL;                                                                  \
+    /* MARK: List Dart lifecycle callbacks */                                                                       \
+                                                                                                                    \
+    [appServiceClient listDartLifecycleCallbacksWithCompletion:^(NSArray<NSString *> *_Nullable setUpAlls,          \
+                                                                 NSArray<NSString *> *_Nullable tearDownAlls,       \
+                                                                 NSError *_Nullable err) {                          \
+      if (err != NULL) {                                                                                            \
+        NSLog(@"listDartLifecycleCallbacks(): failed, err: %@", err);                                               \
+      }                                                                                                             \
+                                                                                                                    \
+      callbacksState = [[NSMutableDictionary alloc] init];                                                          \
+      for (NSString * setUpAll in setUpAlls) {                                                                      \
+        [callbacksState setObject:@NO forKey:setUpAll];                                                             \
+      }                                                                                                             \
+    }];                                                                                                             \
+                                                                                                                    \
+    /* Spin the runloop waiting until the app reports the Dart lifecycle callbacks it contains */                   \
+    while (!callbacksState) {                                                                                       \
+      [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                            \
+    }                                                                                                               \
+    NSLog(@"Got %lu Dart lifecycle callbacks: %@", callbacksState.count, callbacksState);                           \
+                                                                                                                    \
+    /* MARK: List Dart tests */                                                                                     \
+                                                                                                                    \
     [appServiceClient listDartTestsWithCompletion:^(NSArray<NSString *> *_Nullable tests, NSError *_Nullable err) { \
       if (err != NULL) {                                                                                            \
         NSLog(@"listDartTests(): failed, err: %@", err);                                                            \
@@ -56,12 +98,14 @@
     while (!dartTests) {                                                                                            \
       [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                            \
     }                                                                                                               \
-                                                                                                                    \
     NSLog(@"Got %lu Dart tests: %@", dartTests.count, dartTests);                                                   \
+                                                                                                                    \
+    /* MARK: Create tests at runtime */                                                                             \
                                                                                                                     \
     NSMutableArray<NSInvocation *> *invocations = [[NSMutableArray alloc] init];                                    \
                                                                                                                     \
     /**                                                                                                             \
+     *                                                                                                              \
      * Once Dart tests are available, we:                                                                           \
      *                                                                                                              \
      *  Step 1. Dynamically add test case methods that request execution of an individual Dart test file.           \
@@ -73,25 +117,63 @@
       /* Step 1 - dynamically create test cases */                                                                  \
                                                                                                                     \
       IMP implementation = imp_implementationWithBlock(^(id _self) {                                                \
-        [[[XCUIApplication alloc] init] launch];                                                                    \
+        /* Reset server's appReady state, because new app process will be started */                                \
+        appReady = NO;                                                                                              \
                                                                                                                     \
-        __block ObjCRunDartTestResponse *response = NULL;                                                           \
-        [appServiceClient runDartTestWithName:dartTest                                                              \
-                                   completion:^(ObjCRunDartTestResponse *_Nullable r, NSError *_Nullable err) {     \
-                                     if (err != NULL) {                                                             \
-                                       NSLog(@"runDartTestWithName(%@): failed, err: %@", dartTest, err);           \
-                                     }                                                                              \
+        XCUIApplication *app = [[XCUIApplication alloc] init];                                                      \
+        NSDictionary *args = @{@"PATROL_INITIAL_RUN" : @"false"};                                                   \
+        [app setLaunchEnvironment:args];                                                                            \
+        [app launch];                                                                                               \
                                                                                                                     \
-                                     response = r;                                                                  \
-                                   }];                                                                              \
-                                                                                                                    \
-        /* Wait until Dart test finishes */                                                                         \
-        while (!response) {                                                                                         \
+        /* Spin the runloop waiting until the app reports that PatrolAppService is up */                            \
+        while (!appReady) {                                                                                         \
           [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                        \
         }                                                                                                           \
                                                                                                                     \
-        XCTAssertTrue(response.passed, @"%@", response.details);                                                    \
+        __block BOOL callbacksSet = NO;                                                                             \
+        [appServiceClient                                                                                           \
+            setDartLifecycleCallbacksState:callbacksState                                                           \
+                                completion:^(NSError *_Nullable err) {                                              \
+                                  if (err != NULL) {                                                                \
+                                    NSLog(@"setDartLifecycleCallbacksState(): call failed, err: %@", err);          \
+                                  }                                                                                 \
+                                                                                                                    \
+                                  NSLog(@"setDartLifecycleCallbacksState(): call succeeded");                       \
+                                  callbacksSet = YES;                                                               \
+                                }];                                                                                 \
+                                                                                                                    \
+        /*Wait until lifecycle callbacks are set*/                                                                  \
+        while (!callbacksSet) {                                                                                     \
+          [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                        \
+        }                                                                                                           \
+                                                                                                                    \
+        __block ObjCRunDartTestResponse *response = NULL;                                                           \
+        __block NSError *error;                                                                                     \
+        [appServiceClient runDartTest:dartTest                                                                      \
+                           completion:^(ObjCRunDartTestResponse *_Nullable r, NSError *_Nullable err) {             \
+                             NSString *status;                                                                      \
+                             if (err != NULL) {                                                                     \
+                               error = err;                                                                         \
+                               status = @"CRASHED";                                                                 \
+                             } else {                                                                               \
+                               response = r;                                                                        \
+                               status = response.passed ? @"PASSED" : @"FAILED";                                    \
+                             }                                                                                      \
+                                                                                                                    \
+                             NSLog(@"runDartTest(\"%@\"): call finished, test result: %@", dartTest, status);       \
+                           }];                                                                                      \
+                                                                                                                    \
+        /* Wait until Dart test finishes (either fails or passes) or crashes */                                     \
+        while (!response && !error) {                                                                               \
+          [NSRunLoop.currentRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];                        \
+        }                                                                                                           \
+                                                                                                                    \
+        BOOL passed = response ? response.passed : NO;                                                              \
+        NSString *details = response ? response.details : @"(no details - app likely crashed)";                     \
+                                                                                                                    \
+        XCTAssertTrue(passed, @"%@", details);                                                                      \
       });                                                                                                           \
+                                                                                                                    \
       NSString *selectorName = [PatrolUtils createMethodNameFromPatrolGeneratedGroup:dartTest];                     \
       SEL selector = NSSelectorFromString(selectorName);                                                            \
       class_addMethod(self, selector, implementation, "v@:");                                                       \