+++ packages/patrol_cli/CHANGELOG.md
@@ -1,6 +1,20 @@
-## Unreleased
+## 2.4.0-dev.1
 
-- Bump dependencies, populate `topics` in pubspec (#1904)
+- Add support for `patrolSetUpAll` (#1751)
+
+This version requires version 3.1.0-dev1 of `patrol` package.
+
+## 2.3.1+1
+
+- Add screenshots to `pubspec.yaml` (#1917)
+
+## 2.3.1
+
+- Bump dependencies for Flutter 3.16 and Dart 3.2
+- Populate `topics` in pubspec (#1904)
+- Remove warning about `patrol_finders` being split from `patrol`
+- Allow to disable autogenerating shell completion code in `~/.zshrc` (and
+  alike) by setting the `PATROL_NO_COMPLETION` environment variable
 
 ## 2.3.0